Fixed a test
This test isn't running on CI, but Phuong has a PR that fixes that.

I confused the modulus of the field and the order of the curve. This is
still safe since it more hashes than the bitcoin network has ever run to
make our server crash and embarass us a little bit.